Some minor changes from Stefan Allius to make conf and mconf compile under Solaris.

+++ extra/config/mconf.c
@@ -8,6 +8,7 @@
 
 #include <sys/ioctl.h>
 #include <sys/wait.h>
+#include <sys/termios.h>
 #include <ctype.h>
 #include <errno.h>
 #include <fcntl.h>
@@ -443,7 +444,7 @@ static void conf(struct menu *menu)
 		if (!type)
 			continue;
 
-		for (i = 0; input_buf[i] && !isspace(input_buf[i]); i++)
+		for (i = 0; input_buf[i] && !isspace((int)input_buf[i]); i++)
 			;
 		if (i >= sizeof(active_entry))
 			i = sizeof(active_entry) - 1;
